move image widget to widget dir
